requirements.txt
asgiref==3.7.2
Django==4.1.13
django-environ==0.11.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
pymongo==3.12.1
pytz==2023.3.post1
sqlparse==0.2.4
typing_extensions==4.9.0
tzdata==2023.3
